fix bug when file not existed

+++ cmd/gelchain/rebuild.sh
@@ -1,4 +1,6 @@
 cd $GOPATH/src/github.com/green-element-chain/gelchain/cmd/gelchain
-rm -r /usr/bin/gelchain
+if [ -f "/usr/bin/gelchain" ]; then
+        rm -r /usr/bin/gelchain
+fi
 go build
 mv gelchain /usr/bin
